Rename to chart type selector
